Bug 1680669 - part 7: Change the type of `width` and `height` of `PointerEvent` to `double` r=smaug
The values may be used with the coords of `MouseEvent`.  Therefore, they need
to be `double` right now.

Differential Revision: https://phabricator.services.mozilla.com/D222729

+++ widget/MouseEvents.h
@@ -847,8 +847,8 @@ class WidgetPointerEvent : public WidgetMouseEvent {
     return result;
   }
 
-  int32_t mWidth = 1;
-  int32_t mHeight = 1;
+  double mWidth = 1.0;
+  double mHeight = 1.0;
   bool mIsPrimary = true;
   bool mFromTouchEvent = false;
